Added Hazards into the game
