+++ Project_Booster/Assets/Materials/Dummy/Hazard.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b58aa9980fb649d4ea36d6e515fab2e0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
