+++ Project_Booster/Assets/Prefabs/Demo/BoostPanel.prefab
@@ -139,7 +139,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 875f84aa0c596c2499c9d00620a3c997, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speedBoostModifier: 0.05
-  applySpeedBoostTag: Player
+  effectModifier: 0.03
+  colliderTagCheck: Player
   terrainGroupTag: TerrainGroup
   mainCameraTag: MainCamera
