+++ Project_Booster/Assets/Prefabs/Demo/Hazard.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6a4fa33113c325e4bbdf413fb12b29b6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
