+++ Project_Booster/Assets/Scripts/BoostPanel.cs
@@ -6,58 +6,15 @@ using Effects;
 
 namespace Gimmick
 {
-    public class BoostPanel : MonoBehaviour
+    public class BoostPanel : SpawnObj
     {
-        [Header("General Variables")]
-        
-        [Tooltip("The speed buff that the panel grants when touched")]
-        public float speedBoostModifier;
-
-        [Space]
-
-        [Tooltip("The name of the tag to check to apply the boost towards")]
-        public string applySpeedBoostTag;
-
-        [Tooltip("The name of the tag that is associated with the ground/terrain")]
-        public string terrainGroupTag;
-
-        [Tooltip("The name of the tag associated with the main camera")]
-        public string mainCameraTag;
-
-        // Private Variables
-        private static TerrainMovement[] terrainGroups;       // An array of the terrain objects in the level; all boost panels share this information
-        private static CameraMovement mainCamera;           // Ref to the main camera object
-
-        // This finds all of the terrain objects in the level and associates them to the private variable
-        private void Start()
+        protected override void Effect()
         {
-            if (terrainGroups == null)
+            foreach (TerrainMovement currTerrain in terrainGroups)
             {
-                GameObject[] temp = GameObject.FindGameObjectsWithTag(terrainGroupTag);
-                terrainGroups = new TerrainMovement[temp.Length];
-                for (int index = 0; index < terrainGroups.Length; ++index)
-                {
-                    terrainGroups[index] = temp[index].GetComponent<TerrainMovement>();
-                }
+                currTerrain.CurrMoveSpeed += effectModifier;
             }
-
-            if (mainCamera == null)
-            {
-                mainCamera = GameObject.FindGameObjectWithTag(mainCameraTag).GetComponent<CameraMovement>();
-            }
-        }
-
-        // If the specified object touches this, we "mock" the speed up by speeding up the terrain around the player
-        private void OnTriggerEnter(Collider other)
-        {
-            if(other.CompareTag(applySpeedBoostTag))
-            {
-                foreach(TerrainMovement currTerrain in terrainGroups)
-                {
-                    currTerrain.CurrMoveSpeed += speedBoostModifier;
-                }
-                mainCamera.IsBoosting = true;
-            }
+            mainCamera.IsBoosting = true;
         }
     }
 }
