+++ Project_Booster/Assets/Scripts/Hazard.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Gimmick
+{
+    public class Hazard : SpawnObj
+    {
+        protected override void Effect()
+        {
+            print("Hllo");
+        }
+    }
+}
