+++ Project_Booster/Assets/Scripts/Hazard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 168d6f544aea53c47aec9e2a7cd83196
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
