+++ Project_Booster/Assets/Scripts/SpawnObj.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71922fc4a63e2974b8c0d661107340ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
